requirements.txt
discord.py==2.6.4
PyNaCl==1.5.0
python-dotenv==1.0.1
